The main theme of this revision is about making axonal delays specific to particular classes of synapses (excitatory efferents).
utils.c/h
Added EE macro as a wrapper for exit_error() which prints the file, function and line number of the error locus. 

parameters
Added bool initialised to mp, to flag whether the network has already been initialised (with a particular number of layers) which is set at the end of read_parameters.c. This is accompanied by routines for changing all the relevant layer-based vectors of parameters if the number of layers changes after initialisation. 
Changed axonDelay to a bool and added delayEfE, delayElE and delayEI to initialise delays in those types of connection independently of one another. 
Moved the parsing loop for image parameters up to just after the main parsing loop. 
Added some more if statements to handle the Matlab parameters when reading back in parameters.m from a previous simulation. 

spike.c/h
Updated calcMemory for new delay model. Added meanQueue() to tidy the code up for these changes. 
Added recording of the sum of lateral inhibitory synaptic conductances (rec->LsigGI) including printing, reseting and updating routines. N.B. There is a bug in the first recorded value of the training phase. 
Changed create_axons to explicitly pass mp and rewrote for the new synaptic class specific delays. 
Added routines to print out (and later compress) afferent delays.
Removed several old commented out routines. 
Added alternative error catching in enqueue() and dequeue() for when LLVM is used and so asserts can not be used in parallel regions. 

main.c
Now using dlfcn.h to check for access to dynamic libraries. 
Tidied up the output of main and added a lot more checking and information about the compiler, dynamic libraries, OpenMP etc. 
Error messages are now enclosed in "***". 

git-svn-id: https://mac0.cns.ox.ac.uk/svn/SpikeNet/trunk@48 2cedcc20-6066-4341-87df-fb69bdf639a3